# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Hospital Management System Express API with Swagger",
        "description": "This is a Demo app Using JavaScript,NodeJS,ExpressJS,Sequlize with Postgres with Swagger",
        "license": {
            "name": "MIT",
            "url": "https://spdx.org/licenses/MIT.html"
        },
        "contact": {
            "name": "Dhruvik",
            "email": "dhruvik.barvaliya.blackwolve@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": ""
        }
    ],
    "paths": {
...
        "/department": {
            "post": {
                "description": "Create a new Department",
                "tags": [
                    "Department"
                ],
                "responses": {
                    "200": {
                        "description": " The created Department"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/schemas/Department"
                            }
                        }
                    }
                }
            },
            "get": {
                "description": "Lists all the department",
                "tags": [
                    "Department"
                ],
                "responses": {
                    "200": {
                        "description": "The list of the department"
                    }
                }
            }
        },
        "/department/{id}": {
            "get": {
                "description": "Get the Department by id",
                "tags": [
                    "Department"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The Department id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Department response by id"
                    }
                }
            },
            "put": {
                "description": "Update the Department by the id",
                "tags": [
                    "Department"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The Department id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Department was updated"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/schemas/Department"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove the Department by id",
                "tags": [
                    "Department"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The Department id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Department was deleted"
                    }
                }
            }
        },
        "/department/{id}/{status}": {
            "patch": {
                "description": "",
                "tags": [
                    "Department"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The Department id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "description": "Status Of Department",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}